commit
fix uint8_t to uint16_t for freqband variable

+++ RFM69.h
@@ -72,7 +72,7 @@
 #define ISRFM69HW  0
 
 // used function prototypes
-bool RFM69_initialize(uint8_t freqBand, uint8_t nodeID, uint16_t networkID);
+bool RFM69_initialize(uint16_t freqBand, uint8_t nodeID, uint16_t networkID);
 void RFM69_writeReg(uint8_t addr, uint8_t val);
 uint8_t RFM69_readReg(uint8_t addr);
 void RFM69_setAddress(uint8_t addr);
